Curriculum pre req done

+++ app/Http/Controllers/CurriculumCoursesController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\CurriculumCourse;
+use Illuminate\Support\Facades\DB;
+use App\CourseRequisite;
+use App\Http\Resources\CurriculumCourseResource;
 
 class CurriculumCoursesController extends Controller
 {
@@ -35,8 +38,6 @@ class CurriculumCoursesController extends Controller
      */
     public function store(Request $request)
     {
-
-
         $request->validate([
             'course_id' => 'required',
             'curriculum_id' => 'required',
@@ -49,15 +50,49 @@ class CurriculumCoursesController extends Controller
             return response()->json(["message" => "The given data was invalid.","errors" => ["course_id" => ["This course is already added in this curriculum."]]],422);
         }
 
+        //validate pre requisite
+        if (count(request('pre_requisites')) > 0) {
+            foreach (request('pre_requisites') as $pre_req) {
+               if(($pre_req['year_level'] >= request('year_level') && $pre_req['semester'] >= request('semester')) || ($pre_req['year_level'] == request('year_level') && $pre_req['semester'] > request('semester'))) {
+                return response()->json(["message" => "The given data was invalid.","errors" => ["pre_requisites" => ["Invalid pre requisite, make sure you are not creating pre requisite loop. Also, check if the pre requisite is on earlier year level or semester"]]], 422); 
+               }
+            }
+        }
 
-        $curriculum_course = CurriculumCourse::create([
-            'course_id' => request('course_id'),
-            'curriculum_id' => request('curriculum_id'),
-            'year_level' => request('year_level'),
-            'semester' => request('semester')
-        ]);
 
-        return $curriculum_course;
+        DB::beginTransaction();
+
+        try {
+
+            $curriculum_course = CurriculumCourse::create([
+                'course_id' => request('course_id'),
+                'curriculum_id' => request('curriculum_id'),
+                'year_level' => request('year_level'),
+                'semester' => request('semester')
+            ]);
+
+            //save pre requisite
+            if (count(request('pre_requisites')) > 0) {
+                foreach (request('pre_requisites') as $pre_req) {
+                    CourseRequisite::create([
+                        'curriculum_course_id' => $curriculum_course->id,
+                        'type' => 1,
+                        'pre_req_id' => $pre_req["id"]
+                    ]);
+                }
+            }
+            DB::commit();
+            // all good
+            return $curriculum_course;
+
+        } catch (\Exception $e) {
+            DB::rollback();
+            // something went wrong
+            return abort(500, 'Internal Server Error');
+        }
+        
+        
+        //return $request;
     }
 
     /**
@@ -68,7 +103,9 @@ class CurriculumCoursesController extends Controller
      */
     public function show($id)
     {
-        //
+        $curriculum_resource = CurriculumCourse::findOrFail($id);
+
+        return new CurriculumCourseResource($curriculum_resource);
     }
 
     /**
@@ -98,26 +135,61 @@ class CurriculumCoursesController extends Controller
             'semester' => 'required'
         ]);
 
-        // if(CurriculumCourse::where('curriculum_id', '=', request('curriculum_id'))->where('course_id', request('course_id'))->exists()) {
+        //validate pre requisite
+        if (count(request('pre_requisites')) > 0) {
+            foreach (request('pre_requisites') as $pre_req) {
+               if(($pre_req['year_level'] >= request('year_level') && $pre_req['semester'] >= request('semester')) || ($pre_req['year_level'] == request('year_level') && $pre_req['semester'] > request('semester'))) {
+                return response()->json(["message" => "The given data was invalid.","errors" => ["pre_requisites" => ["Invalid pre requisite, make sure you are not creating pre requisite loop. Also, check if the pre requisite is on earlier year level or semester"]]], 422); 
+               }
+            }
+        }
 
-        //     return response()->json(["message" => "The given data was invalid.","errors" => ["course_id" => ["This course is already added in this curriculum."]]],422);
-        // }
 
-
-        // $curriculum_course = CurriculumCourse::create([
-        //     'course_id' => request('course_id'),
-        //     'curriculum_id' => request('curriculum_id'),
-        //     'year_level' => request('year_level'),
-        //     'semester' => request('semester')
-        // ]);
 
         $curriculum_course = CurriculumCourse::findOrFail($id);
+
+
+
+        //check if pre requisite of other courses
+        if(CourseRequisite::where('pre_req_id', '=', $curriculum_course->id)->count() > 0) {
+
+            //check if position change
+            if($curriculum_course->year_level != request('year_level') || $curriculum_course->semester != request('semester')) {
+                return response()->json(["message" => "The given data was invalid.","errors" => ["course_id" => ["Cannot update. This course is pre requisite of other courses"]]], 422);
+            }
+            
+        }
+
+
+
+
         $curriculum_course->course_id = request('course_id');
         $curriculum_course->curriculum_id = request('curriculum_id');
         $curriculum_course->year_level = request('year_level');
         $curriculum_course->semester = request('semester');
 
+
         $curriculum_course->update();
+
+        //delete previous pre requisite(s)
+        foreach ($curriculum_course->courseRequisites as $pre_req) {
+           // $course_req = CourseRequisite::findOrFail($pre_req['pre_req_id']);
+           // $course_req->delete();
+            $pre_req->delete();
+        }
+        
+
+
+        //save pre requisite
+        if (count(request('pre_requisites')) > 0) {
+            foreach (request('pre_requisites') as $pre_req) {
+                CourseRequisite::create([
+                    'curriculum_course_id' => $curriculum_course->id,
+                    'type' => 1,
+                    'pre_req_id' => $pre_req["id"]
+                ]);
+            }
+        }
 
 
         return $curriculum_course;
@@ -132,8 +204,38 @@ class CurriculumCoursesController extends Controller
     public function destroy($id)
     {
         $curriculumCourse = CurriculumCourse::findOrFail($id);
-        $curriculumCourse->delete();
 
-        return $curriculumCourse;
+        DB::beginTransaction();
+
+        try {
+            //validate first before delete
+            if(CourseRequisite::where('pre_req_id', '=', $curriculumCourse->id)->count() > 0) {
+                return abort(422, 'Unprocess Entity');
+            }
+
+
+
+            //delete previous pre requisite(s)
+            foreach ($curriculumCourse->courseRequisites as $pre_req) {
+                $pre_req->delete();
+            }
+
+            $curriculumCourse->delete();
+
+            
+
+            DB::commit();
+            // all good
+
+            return $curriculumCourse;
+        } catch (\Exception $e) {
+            DB::rollback();
+            // something went wrong
+
+            return abort(500, 'Internal Server Error');
+        }
+
+
+        
     }
 }
